requirements.txt
aloe==0.2.0
aloe-django==0.2.0
ansicolors==1.1.8
asgiref==3.2.5
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cssselect==1.1.0
dj-database-url==0.5.0
Django==3.0.4
django-environ==0.4.5
django-heroku==0.3.1
django-nose==1.4.6
djangorestframework==3.11.0
gherkin-official==4.1.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
nose==1.3.7
psycopg2==2.8.4
pylint==2.4.4
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
rq==1.3.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
wrapt==1.11.2
